requirements.txt
asgiref==3.5.0
autopep8==1.6.0
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
dnspython==2.2.0
docker==5.0.3
idna==3.3
importlib-metadata==4.11.1
ipfshttpclient==0.8.0a1
Markdown==3.3.6
multiaddr==0.0.9
netaddr==0.8.0
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
varint==1.0.2
websocket-client==1.2.3
win32-setctime==1.1.0
wrapt==1.13.3
zipp==3.7.0
